hacking: add initial support for firecracker
--I've hacked my way through sandboxCreate up to the
point that we are calling into container create.

So far the only real issue I see is that we want the
sandbox to be running (cal lto the agent) before we create
the initial container.  We may need to split up the create container
to a finer granularity such that the resources can be parsed first
so we can actually start the virtual machine and then call sandbox start
and container create in guest.

There are placeholders for where to make the calls to the API rest
server for firecracker.  So far only the instance-info and initial
firecracker create are needed.

Signed-off-by: Eric Ernst <eric.ernst@intel.com>