hack: remove 9p, use gRPC instead
Add a new gRPC function to send the 'mostly' static
files which were previously proxied in using 9p.  With this in place
we'd be able to boot Kata without *any* 9p.  This is good news for using
hypervisors that don't support 9p, as well as a good configuration to
avoid using 9p.

This won't handle the case where the file is later modified, as we are
creating a copy instead of bindmounting.

While we should trust the host, we may want to make sure the file is a
sane size.

Has dependency on agent implementing the new command.

Signed-off-by: Eric Ernst <eric.ernst@intel.com>